# project-management-system/PMS-Swagger- Documentation.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SriramB/ProjectManagementSystem/1.0.0
  - description: URL for PMS
    url: http://localhost:8080/api/v1
info:
  description: This API is PMS
  version: "1.0.0"
  title: PMS
  contact:
    email: suba@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /DefectHistory/displayDefectHistory:
    get:
      tags:
        - developers
        - analyst
        - manager
        - tester
        - admins
      security:
        - bearerAuth: []
      summary: Search for Defect History List
      operationId: getAllDefectHistory
      description: |
        Search for Defect History List
      responses:
        '200':
          description: List of Defect History
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DefectHistoryList'
        '400':
          description: Invalid Input
  /DefectHistory/getDefectHistoryByDefectId/{defectId}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for Specific Defect History
      operationId: getDefectHistoryByDefectId
      description:
        Search for Specific Defect History
      parameters:
        - in: path
          name: defectId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Specific Defect History
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefectHistoryOne'
        '400':
          description: Invalid Input
  /Defect/displayDefects:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for All Defects
      operationId: getAllDefects
      description:
        Search for All Defects
      responses:
        '200':
          description: Get All Active Defects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefectList'
        '400':
          description: Invalid Input
  /Defect/getDefectByDefectId/{defectId}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for Specific Defect
      operationId: getDefectByDefectId
      description:
        Search for Specific Defect details
      parameters:
        - in: path
          name: defectId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Specific Defect 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefectOne'
        '400':
          description: Invalid Input
  /Defect/getDefectByProjectId/{projectId}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for particular project's Defects
      operationId: getDefectByProjectId
      description:
        Search for particular project's Defects
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of  Defects of particular project 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefectList'
        '400':
          description: Invalid Input
  /Defect/removeDefectByDefectId/{defectId}:
    delete:
      tags:
        - admins
        - tester
      security:
        - bearerAuth: []
      summary: Delete a defect
      operationId: removeDefectByDefectId
      description:
        Deleting a defect will also delete defect history and comments associated with it
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Defect has been deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefectDelete'
        '400':
          description: Invalid Input
  /Defect/addDefect:
    post:
      tags:
        - admins
        - tester
        - manager
      security:
        - bearerAuth: []
      summary: Add New Defect
      operationId: addDefect
      description:
        Adding a new defect will automatically add defect history and comments to store the updations and comments for new Defect added.
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDefectContent'
      responses:
        '200':
          description: New Defect is added
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DefectOne'
        '400':
          description: Invalid Input
  /Defect/updateDefectByDefectId/{defectId}:
    patch:
      tags:
        - admins
        - tester
        - manager
      security:
        - bearerAuth: []
      summary: Updating Defect
      operationId: updateDefect
      description: 
        Updating defect will make some entries in defect history.
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDefectContent'
      responses:
        '200':
          description: Active Defect is updated
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DefectOne'
        '400':
          description: Invalid Input
  /Comment/displayComments:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for All Comments
      operationId: getAllComments
      description:
        Search for All Comments
      responses:
        '200':
          description: Get All Active Comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentList'
        '400':
          description: Invalid Input
  /Comment/getCommentByDefectId/{defectId}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for Specific Comment
      operationId: getCommentByDefectId
      description:
        Search for Specific Comment details
      parameters:
        - in: path
          name: defectId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Specific Comment 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentOne'
        '400':
          description: Invalid Input
  /Comment/addChat/{defectId}:
    patch:
      tags:
        - admins
      security:
        - bearerAuth: []
      summary: Chat in comments
      operationId: addChat
      description:
        Add Chats for Specific Comment description
      parameters:
        - in: path
          name: defectId
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddChatContent'
      responses:
        '200':
          description: Add Chat 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentOne'
        '400':
          description: Invalid Input
...
components:
  securitySchemes:
    bearerAuth:   
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DefectHistoryList:
      # type: object
      properties:
        History of Defects:
          type: array
          items:
              $ref: '#/components/schemas/DefectHistoryContent'
    DefectHistoryOne:
      properties:
        Defect History:
          $ref: '#/components/schemas/DefectHistoryContent'
    DefectHistoryContent:
      type: object
      required:
        - defectId
        - defectStatus
        - defectHistory
      properties:
        defectId:
          type: string
          example: D1
        defectStatus:
          type: string
          example: active
        defectHistory:
          type: array
          items:
            $ref: '#/components/schemas/HistoryContent'
    HistoryContent:
      type: object
      required:
        - eventId
        - status
        - startDate
        - endDate
        - workDoneBy
      properties:
        eventId:
          type: integer
          example: 1
        status:
          type: string
          example: active
        startDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        endDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        workDoneBy:
          type: string
          example: U15
    DefectOne:
      properties:
        Defect :
          $ref: '#/components/schemas/DefectContent'
    DefectList:
      properties:
        List of Defects:
          type: array
          items:
              $ref: '#/components/schemas/DefectContent'
    DefectContent:
      type: object
      required:
        - defectId
        - defectStatus
        - createdDate
        - projectId
        - testCaseId
        - expectedResults
        - actualResults
        - assignedTo
        - createdBy
        - updatedBy
        - lastUpdatedDate
        - bugStatus
        - severity
        - defectType
        - attachmentLinks
      properties:
        defectId:
          type: string
          example: D1
        defectStatus:
          type: string
          example: active
        createdDate:
          type: object
          example: "2020-11-08T18:39:54.304+00:00"
        projectId:
          type: string
          example: P1
        testCaseId : 
          type: string
          example: T2
        expectedResults:
          type: string
          example: Should Display project details of project P1 when p1 is entered as input
        actualResults:
          type: string
          example: Invalid input. No records found for p1.
        assignedTo:
          type: string
          example: U14
        createdBy:
          type: string
          example: U15
        updatedBy:
          type: string
          example: U15
        lastUpdatedDate:
          type: object
          example: "2020-11-08T18:39:54.304+00:00"
        bugStatus:
          type: string
          example: new
        severity:
          type: integer
          example: 1
        defectType: 
          type: string
          example: Requirement
        attachmentLinks:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentLinks'
    AttachmentLinks:
      type: string
      example: "https://www.dropbox.com/home/ScreenShots?preview=actualResult.PNG"
    DefectDelete:
      properties:
        Defect:
          $ref: '#/components/schemas/DeletedString'
    DeletedString:
      type: string
      example: Defect Deleted Successfully
    AddDefectContent:
      type: object
      required:
        - testCaseId
        - expectedResults
        - actualResults
        - assignedTo
        - bugStatus
        - severity
        - defectType
        - attachmentLinks
      properties:
        testCaseId: 
          type: string
          example: T1
        expectedResults:
          type: string
          example: Should Display project details of project P1 when p1 is entered as input
        actualResults:
          type: string
          example: Invalid input. No records found for p1.
        assignedTo:
          type: string
          example: U15
        bugStatus:
          type: string
          example: new
        severity:
          type: integer
          example: 1
        defectType: 
          type: string
          example: Requirement
        attachmentLinks:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentLinks'
    UpdateDefectContent:
      type: object
    CommentList:
      properties:
        List of Comments:
          type: array
          items:
              $ref: '#/components/schemas/CommentContent'
    CommentContent:
      type: object
      required:
        - defectId
        - defectStatus
        - commentDescription
      properties:
        defectId:
          type: string
          example: D1
        defectStatus:
          type: string
          example: active
        commentDescription:
          type: array
          items:
            $ref: '#/components/schemas/ChatContent'
    ChatContent:
      type: object
      required:
        - commentId
        - date
        - repliedBy
        - repliedTo
        - comments
      properties:
        commentId:
          type: string
          example: C1
        date:
          type: object
          example: "2020-11-08T18:39:54.304+00:00"
        repliedBy:
          type: string
          example: U15
        repliedTo:
          type: string
          example: U16
        comments:
          type: string
          example: Get Projects by Id is not working
    CommentOne:
      properties:
        Defect :
          $ref: '#/components/schemas/CommentContent'
    AddChatContent:
      type: object
      required:
        - repliedTo
        - comments
      properties:
        repliedTo:
          type: string
          example: U16
        comments:
          type: string
          example: Bug has been opened
...